🚚 Rename mentions completion
